% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{dk_search}
\alias{dk_search}
\title{d-K Search Algorithm}
\usage{
dk_search(X_array, max_d = 10, n.iter, par.spectrum = c(50),
  par.win = c(1, 0))
}
\arguments{
\item{X_array}{list of data arrays, each of which is organized as No.objects * No.observations * No.segments.}

\item{max_d}{integer, maximal value of \code{d} and \code{K} to be considered}

\item{n.iter}{integer, number of iterations in \code{\link{MIC}} fitting}

\item{par.spectrum}{vector, spectral estimation parameters in the order of \code{c(a, wn, nn)} see \code{\link{spec.parzen}}}

\item{par.win}{vector, epoch smoothing parameters see \code{\link{ep_dismat}}}
}
\value{
A list of objects with the following components:
  \item{\code{d}}{selected d}
  \item{\code{K}}{selected K}
  \item{\code{path}}{searching trajectory}
}
\description{
\code{dk_search} selects the optimal eigen-Laplacian dimensionality (\code{d}) and the number of clusters (\code{K})
  jointly based on model assessment criteria like Bayesian Information Criterion (BIC) and adjusted adherence.
}
\details{
Details of the procedure is available in our manuscript.
}
\examples{
\dontrun{
# An example here

## Time series simulation:
  sim <- MIC_sim(alpha = 0.9,  nsub = 10, fs = 200, segs = 10)


## d,K searching: \\strong{(approx. 9 mins running time)}
  dk_search(sim$Data, max_d = 10, n.iter = 10000, par.spectrum = c(50,50,256))

}

}
\references{
Qian Li, Damla Senturk, Catherine A. Sugar, Shanali Jeste, Charlotte DiStefano, Joel Frohlich, Donatello Telesca
  "\emph{Inferring Brain Signals Synchronicity from a Sample of EEG Readings}".
}
