% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC_helper.R
\name{eigen_lap}
\alias{eigen_lap}
\title{Eigen-Laplacian Dimension Reduction}
\usage{
eigen_lap(X, D)
}
\arguments{
\item{X}{3-d array, suggest \code{diss_array} from \code{\link{ep_dismat}}, symmetric distance matrices stacked into a 3-D array.}

\item{D}{integer, dimension of the output, corresponding to the number of eigenvectors extracted from graph Laplacian.}
}
\value{
A list of objects with the following components:
  \item{\code{eig_data}}{A 3-d array of the eigen-Laplacian representation, No.objects * D * No.epochs}
  \item{\code{eig_value}}{matrix, No.epochs * No.objects, complete eigen-values of graph Laplacian matrices}
}
\description{
\code{eigen_lap} carries out eigen-decomposition on the graph Laplacian matrices, which
  operates on distance matrices and yields low dimensional representations of original objects.
}
\details{
This procedure is a direct implementation of the spectral clustering technique explicated by [Ng 2001].
It is also a preprocessing step for \code{\link{MIC}}, following distance characterization by \code{\link{ep_dismat}}.
}
\references{
Ng, Andrew Y., Michael I. Jordan, and Yair Weiss. \emph{On spectral clustering: Analysis and an algorithm.}
  Advances in neural information processing systems 2 (2002): 849-856.
}
