% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC_helper.R
\name{ep_dismat}
\alias{ep_dismat}
\title{Epoch-wise pairwise Distance Matrices}
\usage{
ep_dismat(X, sf, par.spectrum = c(100, sf/2, 512), window = c(1, 0))
}
\arguments{
\item{X}{3-d array of input, organized as:
No.objects * No.observations * No.segments. Here objects refer to the clustering units;
observations refer to the recorded time series at each segment; and segments are pieces of recordings
that are repeatedly observed/measured with same length.}

\item{sf}{integer, sampling frequency}

\item{par.spectrum}{vector, parameter for \code{\link{spec.parzen}} in the order of \code{c(a, wn, nn)}.}

\item{window}{vector, epoch-smoothing setting in the order of window size and overlap size.
For example, non-smoothing setting is equivalent to \code{window = c(1, 0)}, and \code{window = c(3, 1)}
combines adjacent \strong{3 segnments as} epochs allowing for \strong{1 segment} overlapping in between.}
}
\value{
a list of objects with the following components:
  \item{\code{diss_array}}{3-d array of pairwise distance matrices, No.objects*No.objects*No.epochs}
  \item{\code{ave_spec}}{3-d array of epochwise average spectrum see \code{\link{spec.parzen}}}
  \item{\code{fw}}{vector, frequencies used for spectral estimate}
}
\description{
\code{ep_dismat} calculates strucally smoothed pairwise distances across all objects.
  It is an essential prerequisite step of MIC implementation.
}
\details{
This procedure consists of distance characterization across all objects to be clustered on
  and further smoothing into epochs.

The distance metric is based on the \code{\link{TVD}} between a pair of spectral
  densities, calculated by the Parzen lag window estimator \code{\link{spec.parzen}}.

Epochs are constructed by combining adjacent units (segments), which are further allowed
  to overlap with each other at certain amount.
}
\seealso{
\code{\link{MIC_prep}} for usage
}
