% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC_helper.R
\name{TVD}
\alias{TVD}
\title{Total Variation Distance}
\usage{
TVD(w, f1, f2)
}
\arguments{
\item{w}{A vector of sampled points.}

\item{f1}{Density 1 evaluated at values w.}

\item{f2}{Density 2 evaluated at values w.}
}
\value{
A scalar, in range [0,1] if both f1 and f2 are normalized densities.
}
\description{
\code{TVD} returns the Total Variation Distance between two densities.
}
